+++ backend/sqlite_backend.go
@@ -10,7 +10,7 @@ import (
 )
 
 const (
-	sqliteConnectionStringPrefix = "sqlite:"
+	SqliteConnectionStringPrefix = "sqlite://"
 )
 
 type SqliteBackend struct {
@@ -18,13 +18,13 @@ type SqliteBackend struct {
 	rowReader        RowReader
 }
 
-func NewSqliteBackend(connString string) *SqliteBackend {
+func NewSqliteBackend(connString string) (*SqliteBackend, error) {
 	connString = strings.TrimSpace(connString) // remove any leading or trailing whitespace
-	connString = strings.TrimPrefix(connString, sqliteConnectionStringPrefix)
+	connString = strings.TrimPrefix(connString, SqliteConnectionStringPrefix)
 	return &SqliteBackend{
 		connectionString: connString,
 		rowReader:        newSqliteRowReader(),
-	}
+	}, nil
 }
 
 // Connect implements Backend.