+++ constants/backend.go
@@ -1,9 +1,10 @@
 package constants
 
 const (
-	DuckDBBackendName    = "DuckDB"
-	MySQLBackendName     = "MySQL"
-	PostgresBackendName  = "PostgreSQL"
-	SQLiteBackendName    = "SQLite"
-	SteampipeBackendName = "Steampipe"
+	DuckDBBackendName     = "DuckDB"
+	DuckDBInitBackendName = "DuckDBInit"
+	MySQLBackendName      = "MySQL"
+	PostgresBackendName   = "PostgreSQL"
+	SQLiteBackendName     = "SQLite"
+	SteampipeBackendName  = "Steampipe"
 )