+++ constants/database.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	DuckLakeCatalog         = "tailpipe_ducklake"
+	DuckLakeMetadataCatalog = "__ducklake_metadata_" + DuckLakeCatalog
+)