+++ constants/duckdb_extensions.go
@@ -1,5 +1,4 @@
 package constants
 
-// extend this list to include any additional extensions that need to be installed and loaded
-// into the duckdb backend for Powerpipe and Flowpipe
-var DuckDbExtensions = []string{"json"}
+// DuckDbExtensions contains the standard extensions that we load when loading DuckDB
+var DuckDbExtensions = []string{"inet"}