+++ parse/connection.go
@@ -8,6 +8,7 @@ import (
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/gohcl"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
+	"github.com/turbot/pipe-fittings/v2/funcs"
 	"github.com/turbot/pipe-fittings/v2/hclhelpers"
 	"github.com/turbot/pipe-fittings/v2/modconfig"
 	"github.com/zclconf/go-cty/cty"
@@ -19,7 +20,11 @@ func DecodeConnection(block *hcl.Block) (*modconfig.SteampipeConnection, hcl.Dia
 	if diags.HasErrors() {
 		return nil, diags
 	}
-
+	configPath := block.DefRange.Filename
+	evalCtx := &hcl.EvalContext{
+		Variables: make(map[string]cty.Value),
+		Functions: funcs.ContextFunctions(configPath),
+	}
 	connection := modconfig.NewConnection(block)
 
 	// decode the plugin property
@@ -31,7 +36,7 @@ func DecodeConnection(block *hcl.Block) (*modconfig.SteampipeConnection, hcl.Dia
 
 	if connectionContent.Attributes["type"] != nil {
 		var connectionType string
-		diags = gohcl.DecodeExpression(connectionContent.Attributes["type"].Expr, nil, &connectionType)
+		diags = gohcl.DecodeExpression(connectionContent.Attributes["type"].Expr, evalCtx, &connectionType)
 		if diags.HasErrors() {
 			return nil, diags
 		}
@@ -39,7 +44,7 @@ func DecodeConnection(block *hcl.Block) (*modconfig.SteampipeConnection, hcl.Dia
 	}
 	if connectionContent.Attributes["import_schema"] != nil {
 		var importSchema string
-		diags = gohcl.DecodeExpression(connectionContent.Attributes["import_schema"].Expr, nil, &importSchema)
+		diags = gohcl.DecodeExpression(connectionContent.Attributes["import_schema"].Expr, evalCtx, &importSchema)
 		if diags.HasErrors() {
 			return nil, diags
 		}
@@ -47,7 +52,7 @@ func DecodeConnection(block *hcl.Block) (*modconfig.SteampipeConnection, hcl.Dia
 	}
 	if connectionContent.Attributes["connections"] != nil {
 		var connections []string
-		diags = gohcl.DecodeExpression(connectionContent.Attributes["connections"].Expr, nil, &connections)
+		diags = gohcl.DecodeExpression(connectionContent.Attributes["connections"].Expr, evalCtx, &connections)
 		if diags.HasErrors() {
 			return nil, diags
 		}
@@ -74,7 +79,7 @@ func DecodeConnection(block *hcl.Block) (*modconfig.SteampipeConnection, hcl.Dia
 	}
 
 	// convert the remaining config to a hcl string to pass to the plugin
-	config, moreDiags := hclhelpers.HclBodyToHclString(rest, connectionContent)
+	config, moreDiags := hclhelpers.HclBodyToHclString(rest, connectionContent, evalCtx)
 	if moreDiags.HasErrors() {
 		diags = append(diags, moreDiags...)
 	} else {