+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 Shared Pipes Component
 
+## 2.6.3 [2025-08-19]
+
+* Reverted accidental removal of functions in ociinstaller package.
+
+## 2.6.2 [2025-08-19]
+
+* Fixed issue where a version.json file to the working directory if the plugin binary does not exist. ([#751](https://github.com/turbot/pipe-fittings/issues/751))
+* Add support to use functions in Steampipe connection configs. ([#750](https://github.com/turbot/pipe-fittings/issues/750))
+* Upgraded `hashicorp/go-getter` to remediate high vulnerabilities.
+
+## 2.6.1 [2025-08-01]
+
+* Update ResourceFilter to support parsing of JSON path expressions through the `--where` arg. ([#735](https://github.com/turbot/pipe-fittings/issues/735))
+
 ## 2.6.0 [2025-07-24]
 
 * Compiled with Go 1.24.